Addictions and everything is being retrieved now upon Login Calendar Activity fixeddashboad button problem Renamed Views Addictions Deletion is Dynamic now but ONLY inside the ListViewAdapters, Map Adapters bulbs are working also Addictions are being added dynamically too now Deleting user event/ven/org/artist is a bit more fixed

+++ app/src/main/java/com/bruha/bruha/Adapters/EventListviewAdapter.java
@@ -455,11 +455,13 @@ public class EventListviewAdapter extends BaseSwipeAdapter {
                     @Override
                     public void onClick(View v) {
                       String x=  retrieveMyPHP.deleteUserEvent(MyApplication.userName,event.getEventid());
-                        dbHelper.deleteUserEvent(dbHelper.getWritableDatabase(),event.getEventid());
                         Toast.makeText(mActivity.getApplicationContext(),x,Toast.LENGTH_SHORT).show();
 
-                        mEvents.remove(position);
-                        notifyDataSetChanged();
+                        if(x.equals("Event Deleted!")) {
+                            dbHelper.deleteUserEvent(dbHelper.getWritableDatabase(), event.getEventid());
+                            mEvents.remove(position);
+                            notifyDataSetChanged();
+                        }
                     }
                 });
             }
